chore: Divide components to files
